+++ lib/prawn/swiss_qr_bill/extension.rb
@@ -4,8 +4,8 @@ module Prawn
   module SwissQRBill
     # Extend prawn with *swiss_qr_bill* methods
     module Extension
-      def swiss_qr_bill(data)
-        Prawn::SwissQRBill::Bill.new(self, data).draw
+      def swiss_qr_bill(data, options = {})
+        Prawn::SwissQRBill::Bill.new(self, data, options).draw
       end
 
       def swiss_qr_bill_sections