+++ lib/prawn/swiss_qr_bill/iban.rb
@@ -2,6 +2,8 @@
 
 module Prawn
   module SwissQRBill
+    class InvalidIBANError < StandardError; end
+
     # Check validity of IBAN.
     #
     # NOTE: *For Switzerland only*